Rename to IntelligentCache
