+++ IntelligentCache.Tests/CompositeCacheTests.cs
@@ -1,12 +1,12 @@
 using AutoFixture;
-using IntelligentHack.DistributedCache;
+using IntelligentHack.IntelligentCache;
 using System;
 using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
 using Xunit;
 
-namespace DistributedCache.Tests
+namespace IntelligentCache.Tests
 {
     public class CompositeCacheTests
     {
