+++ IntelligentCache.Tests/MemoryCacheTests.cs
@@ -1,11 +1,11 @@
 using AutoFixture;
-using IntelligentHack.DistributedCache;
+using IntelligentHack.IntelligentCache;
 using System;
 using System.Threading;
 using System.Threading.Tasks;
 using Xunit;
 
-namespace DistributedCache.Tests
+namespace IntelligentCache.Tests
 {
     public class MemoryCacheTests
     {
