+++ IntelligentCache.sln
@@ -3,14 +3,14 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.30320.27
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DistributedCache", "DistributedCache\DistributedCache.csproj", "{0419D13E-5BCC-45D0-91C0-33C20418C4AE}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "IntelligentCache", "IntelligentCache\IntelligentCache.csproj", "{0419D13E-5BCC-45D0-91C0-33C20418C4AE}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{A09FE95C-B99A-436E-B29F-ADA3F24BBA74}"
 	ProjectSection(SolutionItems) = preProject
 		.editorconfig = .editorconfig
 	EndProjectSection
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DistributedCache.Tests", "DistributedCache.Tests\DistributedCache.Tests.csproj", "{DE5B0C27-C362-4EDB-8C88-8973B5E7B3CA}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "IntelligentCache.Tests", "IntelligentCache.Tests\IntelligentCache.Tests.csproj", "{DE5B0C27-C362-4EDB-8C88-8973B5E7B3CA}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
