+++ IntelligentCache/DefaultRedisValueSerializer.cs
@@ -2,7 +2,7 @@ using Newtonsoft.Json;
 using StackExchange.Redis;
 using System;
 
-namespace IntelligentHack.DistributedCache
+namespace IntelligentHack.IntelligentCache
 {
     public sealed class DefaultRedisValueSerializer : IRedisValueSerializer
     {
