+++ IntelligentCache/MemoryCache.cs
@@ -3,7 +3,7 @@ using System.Collections.Concurrent;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace IntelligentHack.DistributedCache
+namespace IntelligentHack.IntelligentCache
 {
     /// <summary>
     /// An implementation of <see cref="ICache" /> that stores values in memory.
